Fixed most problems

+++ src/main/java/com/github/cheesesoftware/PowerfulPerms/database/DBDocument.java
@@ -19,6 +19,6 @@ public class DBDocument {
     }
     
     public int getInt(String key) {
-        return (int)data.get(key);
+        return (int)((long)data.get(key));
     }
 }
